moved tool to tools package. changed coffin state to boolean.
